Move UI init to interface
