+++ TeamProject/GameTechRenderer.h
@@ -22,29 +22,10 @@ namespace NCL {
 			, public GameTechRendererInterface {
 		public:
 			GameTechRenderer();
-
-			struct UIElement {
-				Vector2 position;
-				Vector2 size;
-				Vector4 color;
-				//GLuint* texture;
-				OGLTexture* texture;
-
-				Vector2 GetPosition() { return position; }
-				Vector2 GetSize() { return size; }
-				Vector4 GetColor() { return color; }
-				OGLTexture* GetTexture() { return texture; }
-			};
-
 			~GameTechRenderer();
-
-			//void RenderFrame()	override;
-
-			//Made AddUIElement() public so the pushdown states can use them, is this a problem?
 
 			Mesh* LoadMesh(const std::string& name) override;
 			Texture* LoadTexture(const std::string& name) override;
-			void AddUIElement(Vector2 position, Vector2 size, Vector4 color, OGLTexture* texture = nullptr);
 
 		protected:
 			void NewRenderLines();
@@ -56,10 +37,9 @@ namespace NCL {
 			void RenderShadowMap();
 			void RenderCamera();
 			void RenderSkybox();
-			void InitCrosshair(); //InitCrosshair and RenderCrosshair Ameya added for crosshair
+
 			void RenderDecals();
 			void RenderQuad();
-			//void AddUIElement(Vector2 position, Vector2 size, Vector4 color, OGLTexture* texture = nullptr);
 			void InitUIQuad();
 			void RenderUI();
 
@@ -67,8 +47,6 @@ namespace NCL {
 
 			void SetDebugStringBufferSizes(size_t newVertCount);
 			void SetDebugLineBufferSizes(size_t newVertCount);
-
-			std::vector<UIElement> uiElements;
 
 			std::unique_ptr<OGLShader> uiShader;
 			std::unique_ptr<OGLShader> sceneShader;
