+++ TeamProject/GameTechRendererInterface.cpp
@@ -12,6 +12,7 @@ namespace NCL::CSC8503 {
 		: window(window)
 		, decalSystem(window->GetScreenSize().x, window->GetScreenSize().y)
 	{
+		initUi();
 	}
 
 	void GameTechRendererInterface::collectFrameObjects(GameWorld* world)
@@ -33,5 +34,35 @@ namespace NCL::CSC8503 {
 
 			frameObjects.emplace_back(render);
 		});
+	}
+
+	void GameTechRendererInterface::initUi()
+	{
+		Vector2 screenCenter = Vector2(0.5f, 0.5f);
+		Vector4 crosshairColor = Vector4(1, 1, 1, 1); // White crosshair
+
+		float lineLength = 0.02f; // Length of the crosshair lines
+		float lineThickness = 0.0025f; // Thickness of each line
+		float horizontalLineThickness = 0.0035f;
+		float horizontalLineLength = 0.015f;
+		float gapSize = 0.0005f; // Gap between the lines
+
+
+		// Crosshair
+		// Left line
+		AddUIElement({ Vector2(screenCenter.x - gapSize - lineLength, screenCenter.y),
+			Vector2(horizontalLineLength, horizontalLineThickness), crosshairColor });
+
+		// Right line
+		AddUIElement({ Vector2(screenCenter.x + gapSize + lineLength, screenCenter.y),
+			Vector2(horizontalLineLength, horizontalLineThickness), crosshairColor });
+
+		// Top line
+		AddUIElement({ Vector2(screenCenter.x, screenCenter.y + gapSize + lineLength),
+			Vector2(lineThickness, lineLength), crosshairColor });
+
+		// Bottom line
+		AddUIElement({ Vector2(screenCenter.x, screenCenter.y - gapSize - lineLength),
+			Vector2(lineThickness, lineLength), crosshairColor });
 	}
 }
